Backend/requirement.txt
flask
flask_sqlalchemy
flask_cors
python-dotenv
flask-bcrypt
PyJWT
